// front-caixa/src/types/ImovelData.ts
export interface ImovelData {
  id: string;
  condominio: string;
  areaPrivativa: string;
  areaTerreno: string;
  areaTotal: string;
  bairro: string;
  cidade: string;
  desconto: string;
  descricao: string;
  descricaoDePagamento: string;
  endereco: string;
  foto: string;
  link: string;
  modalidadeVenda: string;
  numeroImovel: string;
  preco: string;
  quarto: string;
  situacao: string;
  tipoImovel: string;
  uf: string;
  valorAvaliacao: string;
  disponivel: boolean;
  aceitaParcelamento: boolean;
  aceitaFinanciamento: boolean;
  aceitaConsorcio: boolean;
  acaoJudicial: boolean;
}


export interface ImovelDataCard {
  id: string;
  numeroImovel: string;
  tipoImovel: string;
  condominio: string;
  cidade: string;
  bairro: string;
  preco: string;
  foto:string;
}



acaoJudicial: false;
aceitaConsorcio: false;
aceitaFinanciamento: false;
aceitaParcelamento: false;
areaPrivativa: "40.05";
areaTerreno: "200.00";
areaTotal: "0.00";
bairro: "BARREIRO";
cidade: "ALAGOINHAS";
condominio: "LOT PORTO FELIZ";
contaId: 0;
desconto: "57.14";
descricao: "a.serv,  WC,  1 sala(s),  sacada,  cozinha,  1 vaga(s) de garagem.";
descricaoDePagamento: "FORMAS DE PAGAMENTO ACEITAS:\r\n Recursos próprios.\r\n Permite utilização de FGTS. Consulte condições e enquadramento.\r\n\r\nREGRAS PARA PAGAMENTO DAS DESPESAS (caso existam):\r\n Condomínio: Sob responsabilidade do comprador, até o limite de 10% em relação ao valor de avaliação do imóvel. A CAIXA realizará o pagamento apenas do valor que exceder o limite de 10% do valor de avaliação.\r\n Tributos: Sob responsabilidade do comprador.";
disponivel: true;
endereco: "RUA G, N. 181 ";
foto: "https://venda-imoveis.caixa.gov.br/fotos/F844442094225521.jpg";
id: "63771976-1643-41cd-89a6-b6c7582f9511";
link: "https://venda-imoveis.caixa.gov.br/sistema/detalhe-imovel.asp?hdnOrigem=index&hdnimovel=8444420942255";
modalidadeVenda: "Venda Direta Online";
numeroImovel: "8444420942255";
preco: "49.294,28";
quarto: "2";
situacao: "Ocupado";
tipoImovel: "Casa";
uf: "BA";
valorAvaliacao: "115.000,00";